check point - azd help completed
